pyproject.toml
[project]
name = "office-earnings"
version = "0.1.0"
description = "This project parses earnings report in PDFs to meaningful actionable spreadsheets"
authors = [
    {name = "ab1ndal",email = "bindal.abhinav@gmail.com"}
]
readme = "README.md"
packages = [{ include = "office_earnings", from = "src" }]
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "openai (>=1.104.0,<2.0.0)",
    "streamlit (>=1.49.1,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pymupdf (>=1.26.4,<2.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
black = "^25.1.0"
pre-commit = "^4.3.0"
pytest = "^8.4.1"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.black]
line-length = 100
target-version = ["py313"]